Move config entry migration helpers (#153)
